requirements.txt
websocket-client~=1.8.0
numpy~=2.0.1
pyaudio~=0.2.14